+++ app/src/main/java/com/punyo/nitechvacancyviewer/ui/navigation/Navigation.kt
@@ -18,12 +18,12 @@ import com.punyo.nitechvacancyviewer.R
 import com.punyo.nitechvacancyviewer.application.GsonInstance
 import com.punyo.nitechvacancyviewer.data.room.model.Room
 import com.punyo.nitechvacancyviewer.data.setting.model.ThemeSettings
+import com.punyo.nitechvacancyviewer.theme.AppTheme
 import com.punyo.nitechvacancyviewer.ui.component.ReservationTableScreen
 import com.punyo.nitechvacancyviewer.ui.initialize.InitializeScreen
 import com.punyo.nitechvacancyviewer.ui.main.MainScreen
 import com.punyo.nitechvacancyviewer.ui.roomvacancy.RoomVacancyScreen
 import com.punyo.nitechvacancyviewer.ui.signin.SignInScreen
-import com.punyo.nitechvacancyviewer.ui.theme.AppTheme
 
 @Composable
 fun MainNavigation(
